Add tabs and "View all snapshots" back button
